in progress; begin implementing pseudocode.
an issue has arisen here;
because our image is centered vertically on the page,
any change in the width will be applied equally
on the left and right sides of the image.

This means that we need to calculate the new width of the image
to be double the distance that the mouse cursor has moved,
but we only want this style of calculation to apply when the image
is horizontally centered on the page.

So, somehow, our resizing logic needs to query the image layout property
that dictates how it is positioned on the page.
I imagine much more difficulty could arise if the image has, for example,
proportional margins set on each side,
so that dimension changes should be applied
20% on one side and 80% on the other.